// examples/client/src/gen7/models/ts/store/GetOrderById.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { Order } from '../Order.ts'

/**
 * @description ID of order that needs to be fetched
 *
 * Format: `int64`
 * @type integer
 */
export type GetOrderByIdPathOrderId = bigint

/**
 * @type object
 */
export type GetOrderByIdStatus200Json = Order

/**
 * @type object
 */
export type GetOrderByIdStatus200Xml = Order

export type GetOrderByIdStatus200 = GetOrderByIdStatus200Json | GetOrderByIdStatus200Xml

/**
 * @type any
 */
export type GetOrderByIdStatus400 = any

/**
 * @type any
 */
export type GetOrderByIdStatus404 = any

/**
 * @type object
 */
export type GetOrderByIdRequestConfig = {
  body?: never
  /**
   * @type object | undefined
   */
  path?: {
    orderId: GetOrderByIdPathOrderId
  }
  query?: never
  headers?: never
  /**
   * @type string
   */
  url: `/store/order/${string}`
}

/**
 * @type object
 */
export type GetOrderByIdResponses = {
  '200': GetOrderByIdStatus200
  '400': GetOrderByIdStatus400
  '404': GetOrderByIdStatus404
}

/**
 * @description Union of all possible responses
 */
export type GetOrderByIdResponse = GetOrderByIdStatus200 | GetOrderByIdStatus400 | GetOrderByIdStatus404